Refactor Classes
